major refactoring to reduce duplicate code
