+++ src/NukiConstants.h
@@ -16,6 +16,11 @@
 
 namespace Nuki {
 
+    const char BLE_ADDRESS_STORE_NAME[]       = "bleAddress";
+    const char SECURITY_PINCODE_STORE_NAME[]  = "securityPinCode";
+    const char SECRET_KEY_STORE_NAME[]        = "secretKeyK";
+    const char AUTH_ID_STORE_NAME[]           = "authorizationId";
+
     enum class DoorSensorState : uint8_t {
         Unavailable       = 0x00,
         Deactivated       = 0x01,
@@ -53,6 +58,68 @@ namespace Nuki {
         CommandWithChallenge          = 1,
         CommandWithChallengeAndAccept = 2,
         CommandWithChallengeAndPin    = 3
+    };
+
+
+    enum class CommandStatus : uint8_t {
+        Complete        = 0x00,
+        Accepted        = 0x01
+    };
+
+    enum class Command : uint16_t {
+        Empty                         = 0x0000,
+        RequestData	                  = 0x0001,
+        PublicKey	                    = 0x0003,
+        Challenge	                    = 0x0004,
+        AuthorizationAuthenticator	  = 0x0005,
+        AuthorizationData	            = 0x0006,
+        AuthorizationId	              = 0x0007,
+        RemoveUserAuthorization	      = 0x0008,
+        RequestAuthorizationEntries	  = 0x0009,
+        AuthorizationEntry	          = 0x000A,
+        AuthorizationDatInvite	      = 0x000B,
+        KeyturnerStates	              = 0x000C,
+        LockAction	                  = 0x000D,
+        Status	                      = 0x000E,
+        MostRecentCommand	            = 0x000F,
+        OpeningsClosingsSummary	      = 0x0010,
+        BatteryReport	                = 0x0011,
+        ErrorReport	                  = 0x0012,
+        SetConfig	                    = 0x0013,
+        RequestConfig	                = 0x0014,
+        Config	                      = 0x0015,
+        SetSecurityPin	              = 0x0019,
+        RequestCalibration	          = 0x001A,
+        RequestReboot	                = 0x001D,
+        AuthorizationIdConfirmation	  = 0x001E,
+        AuthorizationIdInvite	        = 0x001F,
+        VerifySecurityPin	            = 0x0020,
+        UpdateTime	                  = 0x0021,
+        UpdateAuthorization	          = 0x0025,
+        AuthorizationEntryCount	      = 0x0027,
+        RequestLogEntries	            = 0x0031,
+        LogEntry	                    = 0x0032,
+        LogEntryCount	                = 0x0033,
+        EnableLogging	                = 0x0034,
+        SetAdvancedConfig	            = 0x0035,
+        RequestAdvancedConfig	        = 0x0036,
+        AdvancedConfig	              = 0x0037,
+        AddTimeControlEntry	          = 0x0039,
+        TimeControlEntryId	          = 0x003A,
+        RemoveTimeControlEntry	      = 0x003B,
+        RequestTimeControlEntries	    = 0x003C,
+        TimeControlEntryCount	        = 0x003D,
+        TimeControlEntry	            = 0x003E,
+        UpdateTimeControlEntry	      = 0x003F,
+        AddKeypadCode	                = 0x0041,
+        KeypadCodeId	                = 0x0042,
+        RequestKeypadCodes	          = 0x0043,
+        KeypadCodeCount	              = 0x0044,
+        KeypadCode	                  = 0x0045,
+        UpdateKeypadCode	            = 0x0046,
+        RemoveKeypadCode	            = 0x0047,
+        KeypadAction	                = 0x0048,
+        SimpleLockAction	            = 0x0100
     };
 
     enum class TimeZoneId : uint16_t {
